% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.HL.MrootXLS.R
\name{read.HL.MrootXLS}
\alias{read.HL.MrootXLS}
\title{Reads Hydrolight multi-wavelenght EXCELL spreadsheet (Mroot.xls)}
\usage{
read.HL.MrootXLS(file)
}
\arguments{
\item{file}{is the file name of the multi-wavelenght EXCEL spreadsheet (Mroot.xls)}
}
\value{
Returns a long list of parameter output by Hydrolight
}
\description{
Reads Hydrolight multi-wavelenght EXCELL spreadsheet (Mroot.xls)
}
\author{
Simon Bélanger
}
